temp commit for ERMS-1819
continue here: provider generation, platform linking
