+++ app/application.properties
@@ -1,7 +1,7 @@
 #Grails Metadata file
-#Thu Nov 14 15:31:52 CET 2019
-app.buildDate=14.11.2019 15\:31
-app.buildNumber=7469
+#Tue Nov 19 14:00:37 CET 2019
+app.buildDate=19.11.2019 14\:00
+app.buildNumber=7481
 app.buildProfile=development
 app.grails.version=2.5.6
 app.name=laser
