+++ app/deprecated/grails-app/controllers/com/k-int/kbplus/IdentifierController.groovy
