+++ app/deprecated/grails-app/controllers/com/k-int/kbplus/ProcessLoginController.groovy
@@ -14,12 +14,12 @@ class ProcessLoginController {
   def index() {
     return
 
-    log.debug("ProcessLoginController::index() - session = ${request.session.id}");
+      com.k_int.kbplus.ProcessLoginController.log.debug("ProcessLoginController::index() - session = ${request.session.id}");
 
     def ctx = grailsApplication.mainContext
 
     // Check that request comes from 127.0.0.1
-    log.debug("Auth request from ${request.getRemoteAddr()}");
+    com.k_int.kbplus.ProcessLoginController.log.debug("Auth request from ${request.getRemoteAddr()}");
 
     if ( request.getRemoteAddr() == '127.0.0.1' ) {
     }
@@ -36,9 +36,8 @@ class ProcessLoginController {
     }
 
 
-
-    log.debug("remote institution appears to be : ${params.ea_edinaOrgId}");
-    log.debug("remote user appears to be : ${params.ea_edinaUserId}");
+      com.k_int.kbplus.ProcessLoginController.log.debug("remote institution appears to be : ${params.ea_edinaOrgId}");
+      com.k_int.kbplus.ProcessLoginController.log.debug("remote user appears to be : ${params.ea_edinaUserId}");
 
     def response_str = 'NO EA_EXTRA FOUND';
 
@@ -47,16 +46,16 @@ class ProcessLoginController {
         kv.split('=').toList(); map[key] = value != null ? URLDecoder.decode(value) : null; map 
       }
 
-      log.debug("Auth inst = ${map.authInstitutionName}");
-      log.debug("UserId = ${map.eduPersonTargetedID}");
-      log.debug("email = ${map.mail}");
-      log.debug("Inst Addr = ${map.authInstitutionAddress}");
+        com.k_int.kbplus.ProcessLoginController.log.debug("Auth inst = ${map.authInstitutionName}");
+        com.k_int.kbplus.ProcessLoginController.log.debug("UserId = ${map.eduPersonTargetedID}");
+        com.k_int.kbplus.ProcessLoginController.log.debug("email = ${map.mail}");
+        com.k_int.kbplus.ProcessLoginController.log.debug("Inst Addr = ${map.authInstitutionAddress}");
 
       if ( ( map.eduPersonTargetedID != null ) && ( map.eduPersonTargetedID.length() > 0 ) ) {
   
         def user = com.k_int.kbplus.auth.User.findByUsername(map.eduPersonTargetedID)
         if ( !user ) {
-          log.debug("Creating user");
+          com.k_int.kbplus.ProcessLoginController.log.debug("Creating user");
           user = new com.k_int.kbplus.auth.User(username:map.eduPersonTargetedID,
                                                 password:'**',
                                                 enabled:true,
@@ -67,38 +66,38 @@ class ProcessLoginController {
                                                 email:map.mail)
   
           if ( user.save(flush:true) ) {
-            log.debug("Created user, allocating user role");
+            com.k_int.kbplus.ProcessLoginController.log.debug("Created user, allocating user role");
             def userRole = com.k_int.kbplus.auth.Role.findByAuthority('ROLE_USER')
   
             if ( userRole ) {
-              log.debug("looked up user role: ${userRole}");
+              com.k_int.kbplus.ProcessLoginController.log.debug("looked up user role: ${userRole}");
               def new_role_allocation = new com.k_int.kbplus.auth.UserRole(user:user,role:userRole);
     
               if ( new_role_allocation.save(flush:true) ) {
-                log.debug("New role created...");
+                com.k_int.kbplus.ProcessLoginController.log.debug("New role created...");
               }
               else {
                 new_role_allocation.errors.each { e ->
-                  log.error(e);
+                    com.k_int.kbplus.ProcessLoginController.log.error(e);
                 }
               }
             }
             else {
-              log.error("Unable to look up ROLE_USER");
+              com.k_int.kbplus.ProcessLoginController.log.error("Unable to look up ROLE_USER");
             }
 
             // See if we can find the org this user is attached to
             if ( grailsApplication.config.autoAffiliate ) {
               createUserOrgLink(user, map.authInstitutionName, map.shibbScope);
             }
-    
-            log.debug("Done creating user");
+
+              com.k_int.kbplus.ProcessLoginController.log.debug("Done creating user");
           }
         }
         else {
-          log.error("Problem creating user......");
+          com.k_int.kbplus.ProcessLoginController.log.error("Problem creating user......");
           user.errors.each { err ->
-            log.error(err);
+              com.k_int.kbplus.ProcessLoginController.log.error(err);
           }
         }
       
@@ -109,9 +108,9 @@ class ProcessLoginController {
   
         def tok = java.util.UUID.randomUUID().toString()
         ediAuthTokenMap[tok] = map.eduPersonTargetedID
-  
-        log.debug("Setting entry in ediAuthTokenMap to ${tok} = ${map.eduPersonTargetedID}");
-        log.debug(ediAuthTokenMap)
+
+          com.k_int.kbplus.ProcessLoginController.log.debug("Setting entry in ediAuthTokenMap to ${tok} = ${map.eduPersonTargetedID}");
+          com.k_int.kbplus.ProcessLoginController.log.debug(ediAuthTokenMap)
   
         if ( ( params.ea_context ) && ( params.ea_context.trim().length() > 0 ) ) {
           if ( params.ea_context.indexOf('?') > 0 ) {
@@ -131,7 +130,7 @@ class ProcessLoginController {
       }
     }
 
-    log.debug("Rendering processLoginController response, URL will be ${response_str}");
+      com.k_int.kbplus.ProcessLoginController.log.debug("Rendering processLoginController response, URL will be ${response_str}");
 
     // redirect(controller:'home');
     render "${response_str}"
@@ -166,14 +165,14 @@ class ProcessLoginController {
                                                               dateRequested:System.currentTimeMillis(), 
                                                               dateActioned:System.currentTimeMillis())
         if ( !user_org_link.save(flush:true) ) {
-          log.error("Problem saving user org link");
+          com.k_int.kbplus.ProcessLoginController.log.error("Problem saving user org link");
           user_org_link.errors.each { e ->
-            log.error(e);
+              com.k_int.kbplus.ProcessLoginController.log.error(e);
           }
 
         }
         else {
-          log.debug("Linked user with org ${org.id} based on name ${authInstitutionName}");
+          com.k_int.kbplus.ProcessLoginController.log.debug("Linked user with org ${org.id} based on name ${authInstitutionName}");
         }
       }
     }
